backend/pyproject.toml
[project]
name = "travel-builder-api"
version = "0.1.0"
description = "FastAPI backend for Custom Travel Builder"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.32",
  "psycopg[binary]>=3.2.1",
  "pydantic>=2.7.4",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt<4",
  "httpx>=0.27.0",
  "google-generativeai>=0.7.2",
  "openai>=1.12.0",
  "langchain>=1.2.0,<2",
  "langchain-openai>=1.1.0,<2",
  "langchain-google-genai>=4.2.0,<5",
  "langchain-community>=0.4.0,<0.5",
  "langgraph>=1.0.0,<2"
]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["app"]

[tool.uvicorn]
factory = false
